pyproject.toml
[tool.poetry]
name = "fpg_gateway_filterer"
version = "0.1.0"
description = ""
authors = ["W. Ross Morrow <morrowwr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flatten-json = "^0.1.13"
PyYAML = "^6.0"
protobuf = "^3.19.1"
grpcio = "^1.42.0"
protoc-gen-validate = "^0.4.2"
grpcio-reflection = "^1.43.0"
grpc-interceptor = "^0.13.1"
pyjwt = "^2.3.0"
auth0-python = "^3.22.0"
redis = "^4.2.2"
confluent-kafka = "^1.8.2"
ddtrace = "^1.1.0"
JSON-log-formatter = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.3.2"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"


[tool.coverage.paths]
source = ["config"]

[tool.coverage.run]
branch = true
source = ["config"]
omit = []

[tool.coverage.report]
show_missing = true
fail_under = 0 # TODO: increase this as unit tests are added

[tool.black]
line-length = 100
include = 'extproc\/.*\.pyi?$|tests\/.*\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | generated
  )/
)
'''

[tool.isort]
src_paths = ["extproc", "tests"]
profile = "black"
line_length = 80
multi_line_output = 3
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
force_sort_within_sections = true
force_alphabetical_sort_within_sections = true
extend_skip = "generated"
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
